after disconnect move item from request to fetch
